Create Stars_life.py
Nice work :) !, I saw all animations working on glowscript. But as your code is in python, so keep your code in a .py file so that everyone could run it. Also, learn something about a file called 'requirements.txt' which should have details of libraries to be imported for your project.